Simplify blocktool cli structure
