+++ gr-utils/python/blocktool/cli/parseheader.py
@@ -18,7 +18,7 @@
 # the Free Software Foundation, Inc., 51 Franklin Street,
 # Boston, MA 02110-1301, USA.
 #
-""" Module to generate YAML file from the parsed data """
+""" Module to generate parsed header output data """
 
 from __future__ import print_function
 from __future__ import absolute_import
@@ -27,6 +27,8 @@ from __future__ import unicode_literals
 from collections import OrderedDict
 
 import os
+import json
+import logging
 import click
 import yaml
 try:
@@ -34,8 +36,10 @@ try:
 except:
     from yaml import Loader, Dumper
 
-from blocktool.cli.base import run
+from blocktool.cli.base import run, BlockToolException
 from blocktool.core.parseheader import BlockHeaderParser
+
+LOGGER = logging.getLogger(__name__)
 
 ## setup dumper for dumping OrderedDict ##
 _MAPPING_TAG = yaml.resolver.BaseResolver.DEFAULT_MAPPING_TAG
@@ -55,19 +59,43 @@ Dumper.add_representer(OrderedDict, dict_representer)
 Loader.add_constructor(_MAPPING_TAG, dict_constructor)
 
 
-@click.command('makeyaml', short_help='Generate a YAML file from a block header')
+@click.command('parseheader',
+               short_help='Generate the parsed output for the header file or directory in a specified format')
 @click.argument('file-path', nargs=1)
+@click.option('-yaml', '--yaml-confirm', is_flag=True,
+              help='If given, a YAML file alongside a JSON will also be generated')
 def cli(**kwargs):
     """
     \b
     Get parsed output for a header file from GNU Radio module
     """
-    self = BlockHeaderParser(**kwargs)
-    self.cli = True
-    self.yaml_confirm = True
-    click.secho('Header file: {}'.format(self.filename), fg='green')
-    run(self)
-    yaml_generator(self)
+    if os.path.isfile(kwargs['file_path']):
+        self = BlockHeaderParser(**kwargs)
+        self.cli = True
+        self.json_confirm = True
+        click.secho('Header file: {}'.format(self.filename), fg='green')
+        run(self)
+        json_generator(self)
+        if kwargs['yaml_confirm']:
+            self.yaml_confirm = True
+            yaml_generator(self)
+    elif os.path.isdir(kwargs['file_path']):
+        parse_directory(**kwargs)
+    else:
+        raise BlockToolException('Invalid file or directory path.')
+
+
+def json_generator(self):
+    """
+    Generate JSON file for the block header
+    """
+    header_file = self.filename.split('.')[0]
+    block_module = self.modname.split('-')[-1]
+    click.secho('Successfully generated {}_{}.json'.format(
+        block_module, header_file), fg='green')
+    json_file = os.path.join('.', block_module+'_'+header_file + '.json')
+    with open(json_file, 'w') as _file:
+        json.dump(self.parsed_data, _file, indent=4)
 
 
 def yaml_generator(self):
@@ -130,9 +158,10 @@ def yaml_generator(self):
     data['output'] = self.parsed_data['io_signature']['output']
 
     _cpp_templates = [('includes', '#include <gnuradio/{}/{}>'.format(block, self.filename)),
-                  ('declartions', '{}::{}::sptr ${{id}}'.format(block, header)),
-                  ('make', 'this->${{id}} = {}::{}::make({})'.format(block, header, ', '.join(params_list)))
-                  ]
+                      ('declartions', '{}::{}::sptr ${{id}}'.format(block, header)),
+                      ('make', 'this->${{id}} = {}::{}::make({})'.format(
+                          block, header, ', '.join(params_list)))
+                      ]
 
     if self.parsed_data['methods']:
         list_callbacks = []
@@ -162,3 +191,33 @@ def yaml_generator(self):
 
     with open(yml_file, 'w') as yml:
         yaml.dump(data, yml, Dumper=Dumper, default_flow_style=False)
+
+
+def parse_directory(**kwargs):
+    """
+    Get parsed json and yaml output for complete header directory
+    """
+    dir_path = kwargs['file_path']
+    dir_path = os.path.abspath(dir_path)
+    list_header = []
+    dir_name = os.path.basename(dir_path)
+    for _header in os.listdir(dir_path):
+        if _header.endswith('.h') and os.path.isfile(os.path.join(dir_path, _header)):
+            list_header.append(os.path.join(dir_path, _header))
+    list_header = sorted(list_header)
+    for header_path in list_header:
+        kwargs['file_path'] = header_path
+        header = os.path.basename(header_path)
+        try:
+            self = BlockHeaderParser(**kwargs)
+            self.yaml_confirm = True
+            self.json_confirm = True
+            run(self)
+            yaml_generator(self)
+            json_generator(self)
+        except:
+            logging.basicConfig(level=logging.DEBUG,
+                                filename=os.path.join('.', dir_name+'_log.out'))
+            logging.exception(
+                'Log for Exception raised for the header: {}\n'.format(header))
+            click.secho('Parsing unsuccessful: {}'.format(header), fg='yellow')
